Ran Codel Clean up tool

+++ BugTracker/Controllers/CompaniesController.cs
@@ -17,25 +17,19 @@ public class CompaniesController : Controller
     // GET: Companies
     public async Task<IActionResult> Index()
     {
-        return _context.Companies != null ?
-            View(await _context.Companies.ToListAsync()) :
-            Problem("Entity set 'ApplicationDbContext.Companies'  is null.");
+        return _context.Companies != null
+            ? View(await _context.Companies.ToListAsync())
+            : Problem("Entity set 'ApplicationDbContext.Companies'  is null.");
     }
 
     // GET: Companies/Details/5
     public async Task<IActionResult> Details(int? id)
     {
-        if (id == null || _context.Companies == null)
-        {
-            return NotFound();
-        }
+        if (id == null || _context.Companies == null) return NotFound();
 
         var company = await _context.Companies
             .FirstOrDefaultAsync(m => m.Id == id);
-        if (company == null)
-        {
-            return NotFound();
-        }
+        if (company == null) return NotFound();
 
         return View(company);
     }
@@ -59,22 +53,17 @@ public class CompaniesController : Controller
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
+
         return View(company);
     }
 
     // GET: Companies/Edit/5
     public async Task<IActionResult> Edit(int? id)
     {
-        if (id == null || _context.Companies == null)
-        {
-            return NotFound();
-        }
+        if (id == null || _context.Companies == null) return NotFound();
 
         var company = await _context.Companies.FindAsync(id);
-        if (company == null)
-        {
-            return NotFound();
-        }
+        if (company == null) return NotFound();
         return View(company);
     }
 
@@ -85,10 +74,7 @@ public class CompaniesController : Controller
     [ValidateAntiForgeryToken]
     public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Description")] Company company)
     {
-        if (id != company.Id)
-        {
-            return NotFound();
-        }
+        if (id != company.Id) return NotFound();
 
         if (ModelState.IsValid)
         {
@@ -100,51 +86,37 @@ public class CompaniesController : Controller
             catch (DbUpdateConcurrencyException)
             {
                 if (!CompanyExists(company.Id))
-                {
                     return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
+                throw;
             }
+
             return RedirectToAction(nameof(Index));
         }
+
         return View(company);
     }
 
     // GET: Companies/Delete/5
     public async Task<IActionResult> Delete(int? id)
     {
-        if (id == null || _context.Companies == null)
-        {
-            return NotFound();
-        }
+        if (id == null || _context.Companies == null) return NotFound();
 
         var company = await _context.Companies
             .FirstOrDefaultAsync(m => m.Id == id);
-        if (company == null)
-        {
-            return NotFound();
-        }
+        if (company == null) return NotFound();
 
         return View(company);
     }
 
     // POST: Companies/Delete/5
-    [HttpPost, ActionName("Delete")]
+    [HttpPost]
+    [ActionName("Delete")]
     [ValidateAntiForgeryToken]
     public async Task<IActionResult> DeleteConfirmed(int id)
     {
-        if (_context.Companies == null)
-        {
-            return Problem("Entity set 'ApplicationDbContext.Companies'  is null.");
-        }
+        if (_context.Companies == null) return Problem("Entity set 'ApplicationDbContext.Companies'  is null.");
         var company = await _context.Companies.FindAsync(id);
-        if (company != null)
-        {
-            _context.Companies.Remove(company);
-        }
+        if (company != null) _context.Companies.Remove(company);
 
         await _context.SaveChangesAsync();
         return RedirectToAction(nameof(Index));
